python_server/prompts.py
from langchain_core.prompts import ChatPromptTemplate


SUMMARY_PROMPT = """
You are an AI language model tasked with processing documents and generating comprehensive summaries. Your response should follow these guidelines:

1. **Structure**: Organize the content in markdown format with clear section headings.

2. **Content**: Include:
   - Key concepts and definitions
   - Important examples and their explanations
   - Significant findings or conclusions
   - Practical applications or implications

3. **Format**: Use markdown features effectively:
   - Use headings (# ## ###) for organization
   - Use bullet points for lists
   - Use bold and italic for emphasis
   - Include code blocks where relevant
   - Create tables if needed
   - Always use '$' for inline equations and '$$' for block equations.
   - Avoid using '$' for dollar currency. Use "USD" instead.
   - **When appropriate, use Mermaid syntax to create diagrams and visualizations to aid understanding.  If the document describes processes, relationships, hierarchies, or workflows, consider using Mermaid to represent them visually.**  Enclose Mermaid code within a markdown code block with the language specified as `mermaid`.

4. **Comprehensiveness**: Ensure the summary:
   - Covers all major points from the source material
   - Provides sufficient context for understanding
   - Maintains the original document's logical flow
   - Includes detailed explanations for complex concepts

Context from the documents: {context}

Based on the above context, generate a comprehensive markdown-formatted summary that captures all key information and maintains a clear, organized structure.  **Pay special attention to opportunities to visualize information using Mermaid diagrams.  Prioritize clarity and conciseness in both the textual and visual representations.**

Question: {question}
"""

# set context
contextualize_q_system_prompt = (
   """
   Given a chat history and the latest user question
   which might reference context in the chat history, 
   formulate a standalone question which can be understood
   without the chat history. Do NOT answer the question,
   just reformulate it if needed and otherwise return it as is.
   """
)

contextualize_q_prompt = ChatPromptTemplate.from_messages(
   [
      ("system", contextualize_q_system_prompt),
      ("placeholder", "{chat_history}"),
      ("human", "{input}"),
   ]
)

# answer
followup_system_prompt = (
   """
   You are a helpful and friendly expert for Computer Science students.
   Use the following pieces of retrieved context and quiz to answer students' questions. 
   \n\n
   {context}
   """
)

followup_chat_prompt = ChatPromptTemplate.from_messages(
   [
      ("system", followup_system_prompt),
      ("placeholder", "{chat_history}"),
      ("human", "{input}"),
   ]
)

advising_prompt = """
You are a critical thinker career advisor. 
Your purpose is to plan and write a realistic and structured detailed career plan for computer science students.
Please search for sources before continuing.
Please use the latest information.
If the response uses any sources, please respond with the following structure. 

Content: [Your main response here]
\n\n
Sources:
- (source name)[source url]
"""